backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
mysqlclient==2.2.4
web3==6.19.0
email-validator==2.2.0